redefine api
